MRVoxels: split MRFloatGrid.cpp by OpenVDB tool family
MRFloatGrid.cpp is the longest-pole translation unit in the MRVoxels
build (~50 s, dominated by OpenVDB template code generation). Split it
into independent .cpp files grouped by OpenVDB tool family so that codegen
work compiles in parallel instead of inside one serial TU:

- MRFloatGrid.cpp           - FloatGrid wrapper class + light per-voxel accessors
- MRFloatGridResample.cpp   - resampled(), cropped() (GridTransformer / resampleToMatch)
- MRFloatGridFilter.cpp     - gaussianFilter(ed)(), countVoxelsWithValue*() (tools::Filter / foreach)
- MRFloatGridArithmetic.cpp - CSG operators +=, +, -=, -, *=, * (csgUnion / Difference / Intersection)

Function bodies are unchanged - this is a pure source reorganization.
New files are registered in MRVoxels.vcxproj; the CMake build globs *.cpp.
