+++ source/MRVoxels/MRFloatGridResample.cpp
@@ -1,0 +1,104 @@
+#include "MRFloatGrid.h"
+#include "MRVDBFloatGrid.h"
+#include "MRVDBProgressInterrupter.h"
+
+#include "MRMesh/MRVector3.h"
+#include "MRMesh/MRBox.h"
+#include "MRMesh/MRTimer.h"
+#include "MRPch/MRSpdlog.h"
+
+namespace MR
+{
+
+FloatGrid resampled( const FloatGrid& grid, const Vector3f& voxelScale, ProgressCallback cb )
+{
+    if ( !grid )
+        return {};
+    const openvdb::FloatGrid & grid_ = *grid;
+    MR_TIMER;
+    openvdb::FloatGrid::Ptr dest = openvdb::FloatGrid::create( grid->background() );
+    dest->setGridClass( grid->getGridClass() );
+    openvdb::Mat4R transform;
+    transform.setToScale( openvdb::Vec3R{ voxelScale.x,voxelScale.y,voxelScale.z } );
+    dest->setTransform( openvdb::math::Transform::createLinearTransform( transform ) ); // org voxel size is 1.0f
+
+    // just grows to 100%
+    // first grows fast, then slower
+    ProgressCallback dummyProgressCb;
+    float i = 1.0f;
+    if ( cb )
+        dummyProgressCb = [&] ( float )->bool
+    {
+        i += 1e-4f;
+        return cb( 1.0f - 1.0f / std::sqrt( i ) );
+    };
+
+    ProgressInterrupter interrupter( dummyProgressCb );
+    // openvdb::util::NullInterrupter template argument to avoid tbb inconsistency
+
+    // the following piece of code is taken from `openvdb::resampleToMatch` and slightly rewritten to take into account
+    // the specific properties of the usage of OpenVdb in our software
+    bool failed = true;
+
+    // the level set is processed differently due to the potential narrowness of the band near ISO value. It could be just a 1 voxel, in which case
+    // the usual resampling will introduce artifacts and change the topology of ISO surface
+    if ( dest->getGridClass() == openvdb::GRID_LEVEL_SET )
+    {
+        try {
+            // unlike `openvdb::resampleToMatch`, the size of the voxel for the grid is always 1, the true size of the voxel is stored
+            // in volume wrapper
+            dest = openvdb::tools::doLevelSetRebuild( grid_, 0.f, 1, 1, &dest->constTransform(), &interrupter );
+            failed = false;
+        }
+        catch( std::exception& e )
+        {
+            spdlog::warn( "The input grid is classified as a level set, but it has a value type that is not supported by the level set rebuild tool: {}", e.what() );
+        }
+    }
+    // in case of a volume created in "unsigned mode", which is not supported by OpenVdb as level set but still used as such, the result is empty
+    // we detect this case and do resampling
+    if ( failed || dest->evalActiveVoxelDim().asVec3I().product() == 0 )
+    {
+        openvdb::tools::doResampleToMatch<openvdb::tools::BoxSampler, openvdb::util::NullInterrupter>( grid_, *dest, interrupter );
+    }
+
+    if ( interrupter.getWasInterrupted() )
+        return {};
+    // restore normal scale
+    dest->setTransform( openvdb::math::Transform::createLinearTransform( 1.0f ) );
+
+    return MakeFloatGrid( std::move( dest ) );
+}
+
+FloatGrid resampled( const FloatGrid& grid, float voxelScale, ProgressCallback cb )
+{
+    return resampled( grid, Vector3f::diagonal( voxelScale ), cb );
+}
+
+FloatGrid cropped( const FloatGrid& grid, const Box3i& box, ProgressCallback cb )
+{
+    if ( !grid )
+        return {};
+    const openvdb::FloatGrid& grid_ = *grid;
+    MR_TIMER;
+    openvdb::FloatGrid::Ptr dest = openvdb::FloatGrid::create( grid_.background() );
+    dest->setGridClass( grid_.getGridClass() );
+    auto dstAcc = dest->getAccessor();
+    auto srcAcc = grid_.getConstAccessor();
+    size_t pgCounter = 0;
+    size_t volume = size_t( width( box ) ) * height( box ) * depth( box );
+    for ( int z = box.min.z; z < box.max.z; ++z )
+    for ( int y = box.min.y; y < box.max.y; ++y )
+    for ( int x = box.min.x; x < box.max.x; ++x )
+    {
+        openvdb::Coord srcCoord( x, y, z );
+        openvdb::Coord dstCoord( x - box.min.x, y - box.min.y, z - box.min.z );
+        dstAcc.setValue( dstCoord, srcAcc.getValue( srcCoord ) );
+        if ( cb && ( ++pgCounter ) % 1024 == 0 && !cb( float( pgCounter ) / float( volume ) ) )
+            return {};
+    }
+    dest->pruneGrid();
+    return MakeFloatGrid( std::move( dest ) );
+}
+
+} //namespace MR
